refactored models into sub module
