+++ app/models/__init__.py
@@ -1,0 +1,7 @@
+from .user import User as User
+from .note import Note as Note
+from .file import File as File
+from .download import Download as Download
+from .upload import Upload as Upload
+from .bookmark import Bookmark as Bookmark
+from .group import Group as Group
